feat: renderer cache

+++ wasm-src/renderer/gpu/shaders/tile.vs.glsl
@@ -8,14 +8,14 @@ R"(
   in highp vec2 a_position;
   in lowp uvec4 a_color;
   in lowp vec2 a_tex_coord;
-  in lowp vec2 a_tex_coord_curves;
+  in highp vec2 a_tex_coord_curves;
   in highp uint a_attr_1;
   in highp uint a_attr_2;
   in highp uint a_attr_3;
 
   out lowp vec4 v_color;
   out lowp vec2 v_tex_coord;
-  out lowp vec2 v_tex_coord_curves;
+  out highp vec2 v_tex_coord_curves;
 
   flat out highp uint v_attr_1;
   flat out highp uint v_attr_2;
